requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
chromadb==0.4.22
ollama==0.1.6 